pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "vla0-trl"
version = "0.0.2"
description = "Unofficial reimplementation of VLA-0 using TRL's SFTTrainer."
readme = "README.md"
requires-python = ">=3.10"
license = "Apache-2.0"
authors = [
    {name = "Ankit Goyal"},
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "torch>=2.0.0",
    "transformers>=4.40.0",
    "trl>=0.8.0",
    "qwen-vl-utils>=0.0.2",
    # "lerobot @ git+https://github.com/huggingface/lerobot.git@f39652707caed42a7cd5ab36066da5663b9565eb", # commented out this because of too complex dependencies. Install manually afterward!
    "datasets>=2.19.0,<4.0.0",  # lerobot compatibility
    "einops>=0.7.0",
    "Pillow>=10.0.0",
    "kernels",
    "wandb",
]

[project.optional-dependencies]
eval = [
    "mujoco==3.3.2", # IMPORTANT: rendering bug after 3.3.2 version. https://github.com/Lifelong-Robot-Learning/LIBERO/issues/88
    "numpy==1.26.4",
    "libero @ git+https://github.com/MilkClouds/LIBERO.git@minor-fix",  # fork with relaxed deps & modern packaging
    "robosuite",
]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
]

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/rv_train", "src/rv_eval", "src/libero_vqa"]

[tool.ruff]
line-length = 119

[tool.pytest.ini_options]
testpaths = ["scripts"]
python_files = ["test_*.py", "verify_*.py"]
filterwarnings = [
    "ignore::UserWarning",
    "ignore::DeprecationWarning",
]
